requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
cfgv==3.3.1
distlib==0.3.6
Django==3.2.16
django-environ==0.9.0
django-statsd-mozilla==0.4.0
django-tenant-users==1.1.1
django-tenants==3.4.7
django_debug_toolbar==3.8.1
filelock==3.9.0
identify==2.5.12
nodeenv==1.7.0
orjson==3.8.5
Pillow==9.3.0
platformdirs==2.6.2
pre-commit==2.21.0
psycopg2-binary==2.9.5
pytz==2022.7
PyYAML==6.0
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
statsd==4.0.0
virtualenv==20.17.1
zxcvbn==4.4.28
